refactor: move GeoJSON computation server-side in getCitiesGeoJson
Moves the city GeoJSON feature building from a client-side useMemo in
GlobalMap to a server-side getCitiesGeoJson function in events.ts,
reducing client bundle size and aligning with the Astro data-fetching pattern.
